ddr-compare/requirements.txt
pandas>=2.2.0
pyodbc>=4.0.39
databricks-sql-connector>=4.0.0
